added Interface and Clip, Clip-SourceDLine combi implementations
